requirements.txt
absl-py==1.3.0
aed-utilities==0.5.6
annotated-types==0.5.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Augmentor==0.2.12
backcall==0.2.0
bcchapi==1.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
blis==0.7.10
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.9
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
confection==0.1.3
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.14
entrypoints==0.4
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.6.0/es_core_news_md-3.6.0-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.13.0
fire==0.4.0
flatbuffers==24.3.25
fonttools==4.33.3
fsspec==2023.10.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.51.1
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
imageio==2.34.0
imbalanced-learn==0.10.0
inflect==5.6.0
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.6.0
jupyter-client==7.3.1
jupyter-core==4.10.0
kaggle==1.6.3
keras==3.1.1
kiwisolver==1.4.2
langcodes==3.3.0
libclang==14.0.6
lightgbm==3.3.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
missingno==0.5.1
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
mysqlclient==2.1.0
namex==0.0.7
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.2
nltk==3.8
numpy==1.24.4
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.11.0
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathy==0.10.2
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.10.0
plotly-express==0.4.1
preshed==3.0.9
prompt-toolkit==3.0.29
protobuf==4.25.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.12.0
pyparsing==3.0.8
pyreadline3==3.4.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==8.0.1
pytz==2022.1
pywin32==304
PyYAML==6.0.1
pyzmq==22.3.0
randomname==0.1.5
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.9
safetensors==0.4.0
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
selenium==4.14.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlacodegen==2.3.0
SQLAlchemy==1.4.37
srsly==2.4.7
stack-data==0.2.0
statsmodels==0.13.2
sympy==1.12
tenacity==8.0.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.15.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.29.0
termcolor==1.1.0
text-unidecode==1.3
tf-keras-vis==0.8.7
thinc==8.1.12
threadpoolctl==3.1.0
tokenizers==0.15.0
torch==2.1.0
torchvision==0.16.0
tornado==6.1
tqdm==4.64.1
traitlets==5.2.0
transformers==4.35.2
trio==0.22.2
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==1.26.13
validators==0.22.0
wasabi==1.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1

